+++ test/run_all_versions.sh
@@ -147,6 +147,7 @@ SUITES=(
 	native_fastdecode
 	native_fetch_bigcap
 	native_fetch_cache
+	native_fetch_coalesce
 	native_fetch_group_memo
 	native_fetch_interrupt
 	native_fetch_position